fixed weird issues with kill process added waitpid() to kill zombie processes added a sleep() from cancel_menu return removed handler(FAKE_SIGNAL, clock), the program got stuck because of it could not add more than 1 alarm

+++ src/clock/clock.c
@@ -71,23 +71,17 @@ alarm_t create_alarm(alarm_clock_t* clock, time_t time, int difference)
 {
   alarm_t alarm;
   alarm.time = time;
-
   signal(SIGCHLD, (void (*)(int))handler);
-  handler(FAKE_SIGNAL, clock);
-
+  //handler(FAKE_SIGNAL, clock);
   pid_t pid = fork();  
-  
   if (pid == 0) {
     /* child process */
     sleep(difference);
     execl("/bin/mpg123", "/bin/mpg123", "-q" ,"./audio/alarm.mp3", NULL); //TODO make user choose alarm sound.
     exit(EXIT_SUCCESS);
   } 
-
   alarm.pid = pid;
   push(clock, alarm);
-  
- 
   return alarm;
 }
 
@@ -100,31 +94,37 @@ alarm_t create_alarm(alarm_clock_t* clock, time_t time, int difference)
  */
 alarm_t remove_alarm(alarm_clock_t* clock, int index) 
 {
-  alarm_t alarm = clock->alarms[index]; // get alarm we want to remove
+  if(index >= 0){
+    alarm_t alarm = clock->alarms[index]; // get alarm we want to remove
 
 
-  // create temp array with same size as before
-  alarm_t * updated_array = malloc(clock->capacity * sizeof(alarm_t));
+    // create temp array with same size as before
+    alarm_t * updated_array = malloc(clock->capacity * sizeof(alarm_t));
 
-  // move the old array from start (clock->alamrs pointer starts at begining of array)
-  // until it reaches the index.
-  memmove(
-    updated_array,
-    clock->alarms,
-    (index+1)*sizeof(alarm_t)
-  );
+    // move the old array from start (clock->alamrs pointer starts at begining of array)
+    // until it reaches the index.
+    memmove(
+      updated_array,
+      clock->alarms,
+      (index+1)*sizeof(alarm_t)
+    );
 
-  // oppsite here, move last part of array into updated array.
-  memmove(
-    updated_array+index,
-    clock->alarms+(index+1),
-    (clock->capacity - index)*sizeof(alarm_t)
-  );
+    // oppsite here, move last part of array into updated array.
+    memmove(
+      updated_array+index,
+      clock->alarms+(index+1),
+      (clock->capacity - index)*sizeof(alarm_t)
+    );
 
-  free(clock->alarms);  // free old array.
-  clock->alarms = updated_array; // assign new array.
-  clock->length -= 1;
-  kill(alarm.pid, SIGKILL);     // kill alarm process of the child we removed.
+    free(clock->alarms);  // free old array.
+    clock->alarms = updated_array; // assign new array.
+    clock->length -= 1;
+    kill(alarm.pid, SIGKILL);     // kill alarm process of the child we removed.
+    printf("%s\n", "hi");
+    printf("%d\n", clock->length);
+    return alarm;
+  }
+  alarm_t alarm = clock->alarms[0];
   return alarm;
 }
